+++ app-includes/delete-family.php
@@ -72,23 +72,37 @@ echo cota_page_header();
 	<h2>Delete Family</h2>
 	<form class="cota-family-delete" action="delete-family-form-handler.php" method="post">
 		<input type="hidden" name="family_id" value="<?php echo $family['id']; ?>">
-		<label>Family Name</label>
-		<input type="text" name="familyname" value="<?php echo htmlspecialchars( $family['familyname'] ); ?>" readonly>
-		<label>Address</label>
-		<input type="text" name="address" value="<?php echo htmlspecialchars( $family['address'] ); ?>" readonly>
-		<label>City</label>
-		<input type="text" name="city" value="<?php echo htmlspecialchars( $family['city'] ); ?>" readonly>
-		<label>State</label>
-		<input type="text" name="state" value="<?php echo htmlspecialchars( $family['state'] ); ?>" readonly>
-		<label>Zip</label>
-		<input type="text" name="zip" value="<?php echo htmlspecialchars( $family['zip'] ); ?>" readonly>
-		<label>Home Phone</label>
-		<input type="text" name="homephone" value="<?php echo htmlspecialchars( $family['homephone'] ); ?>" readonly>
+		<div class="row g-3">
+			<div class="col-12 col-md-6">
+				<label class="form-label fw-bold">Family Name</label>
+				<input class="form-control" type="text" name="familyname" value="<?php echo htmlspecialchars( $family['familyname'] ); ?>" readonly>
+			</div>
+			<div class="col-12 col-md-6">
+				<label class="form-label fw-bold">Address</label>
+				<input class="form-control" type="text" name="address" value="<?php echo htmlspecialchars( $family['address'] ); ?>" readonly>
+			</div>
+			<div class="col-12 col-md-4">
+				<label class="form-label fw-bold">City</label>
+				<input class="form-control" type="text" name="city" value="<?php echo htmlspecialchars( $family['city'] ); ?>" readonly>
+			</div>
+			<div class="col-12 col-md-4">
+				<label class="form-label fw-bold">State</label>
+				<input class="form-control" type="text" name="state" value="<?php echo htmlspecialchars( $family['state'] ); ?>" readonly>
+			</div>
+			<div class="col-12 col-md-4">
+				<label class="form-label fw-bold">Zip</label>
+				<input class="form-control" type="text" name="zip" value="<?php echo htmlspecialchars( $family['zip'] ); ?>" readonly>
+			</div>
+			<div class="col-12">
+				<label class="form-label fw-bold">Home Phone</label>
+				<input class="form-control" type="text" name="homephone" value="<?php echo htmlspecialchars( $family['homephone'] ); ?>" readonly>
+			</div>
+		</div>
 
-		<button class="delall" type="submit" name="delall" >Delete Entire Family From Directory</button>
+		<button class="delall mt-3" type="submit" name="delall" >Delete Entire Family From Directory</button>
 
-		<h3>Family Members</h3>
-		<div id="members-delete">
+		<h3 class="mt-4">Family Members</h3>
+		<div id="members-delete" class="member-list">
 			<?php
 			$first = true;
 			while ( $member = $members->fetch_assoc() ) :
@@ -108,20 +122,20 @@ echo cota_page_header();
 endif;
 				?>
 			<div class="member-row">
-				<input type="checkbox" name="delete_member[]" value="<?php echo $member['id']; ?>">
-				<input type="text" name="members[first_name][]" value="<?php echo htmlspecialchars( $member['first_name'] ); ?>" readonly>
-				<input type="text" name="members[last_name][]" value="<?php echo ! empty( $member['last_name'] ) ? htmlspecialchars( $member['last_name'] ) : htmlspecialchars( $family['familyname'] ?? '' ); ?>" readonly>
-				<input type="text" name="members[cell_phone][]" value="<?php echo htmlspecialchars( $member['cell_phone'] ); ?>" readonly>
-				<input type="email" name="members[email][]" value="<?php echo htmlspecialchars( $member['email'] ); ?>" readonly>
-				<input type="date" name="members[birthday][]" value="<?php echo htmlspecialchars( $member['birthday'] ); ?>" readonly>
-				<input type="date" name="members[baptism][]" value="<?php echo htmlspecialchars( $member['baptism'] ); ?>" readonly>
+				<input class="form-check-input" type="checkbox" name="delete_member[]" value="<?php echo $member['id']; ?>">
+				<input class="form-control" type="text" name="members[first_name][]" value="<?php echo htmlspecialchars( $member['first_name'] ); ?>" readonly>
+				<input class="form-control" type="text" name="members[last_name][]" value="<?php echo ! empty( $member['last_name'] ) ? htmlspecialchars( $member['last_name'] ) : htmlspecialchars( $family['familyname'] ?? '' ); ?>" readonly>
+				<input class="form-control" type="text" name="members[cell_phone][]" value="<?php echo htmlspecialchars( $member['cell_phone'] ); ?>" readonly>
+				<input class="form-control" type="email" name="members[email][]" value="<?php echo htmlspecialchars( $member['email'] ); ?>" readonly>
+				<input class="form-control" type="date" name="members[birthday][]" value="<?php echo htmlspecialchars( $member['birthday'] ); ?>" readonly>
+				<input class="form-control" type="date" name="members[baptism][]" value="<?php echo htmlspecialchars( $member['baptism'] ); ?>" readonly>
 				<input type="hidden" name="members[id][]" value="<?php echo $member['id']; ?>">
 
 			</div>
 			<?php endwhile; ?>
 		</div>
 
-		<button class="delselected" type="submit" name="delselected">Delete Selected Members From Family</button>
+		<button class="delselected mt-3" type="submit" name="delselected">Delete Selected Members From Family</button>
 	</form>
 	</div>
 </body>